Move Jinja2Blueprint to tools.templating
