+++ pipelex/pipe_operators/pipe_llm_factory.py
@@ -5,7 +5,6 @@ from typing_extensions import Self, override
 
 from pipelex.cogt.llm.llm_models.llm_setting import LLMSettingChoices, LLMSettingOrPresetId
 from pipelex.core.concepts.concept_factory import ConceptFactory
-from pipelex.core.concepts.concept_native import NativeConceptEnum
 from pipelex.core.pipes.pipe_blueprint import PipeBlueprint
 from pipelex.core.pipes.pipe_factory import PipeFactoryProtocol
 from pipelex.core.pipes.pipe_input_spec import InputRequirementBlueprint
@@ -14,8 +13,8 @@ from pipelex.core.pipes.pipe_run_params import make_output_multiplicity
 from pipelex.exceptions import PipeDefinitionError
 from pipelex.hub import get_concept_provider, get_optional_domain
 from pipelex.pipe_operators.llm_prompt_blueprint import LLMPromptBlueprint
-from pipelex.pipe_operators.pipe_jinja2_factory import Jinja2Blueprint, PipeJinja2Blueprint, PipeJinja2Factory
 from pipelex.pipe_operators.pipe_llm import PipeLLM, StructuringMethod
+from pipelex.tools.templating.jinja2_blueprint import Jinja2Blueprint
 from pipelex.tools.templating.jinja2_errors import Jinja2TemplateError
 from pipelex.tools.templating.template_provider_abstract import TemplateNotFoundError
 from pipelex.tools.typing.validation_utils import has_more_than_one_among_attributes_from_lists
